WifiInstaller: add permission for access downloaded files
The previous fix for the security vulnerability b/33178389
add the code to drop the downloaded installation file but
forget to update the manifest for additional permission.  This
omission was due to that the original fix was merged from
mnc-dev, which doesn't have such permission requirement.

So add the required permission for deleting downloaded
installation file.

Bug: 33178389
Test: Verify WifiInstaller works and not crashing due permission
      denied exception

Change-Id: I987ad73dde0e3cc5b920cd08830dd065c47be402