Limit cert file reading to 10MiB.
We read the entire cert file into memory when installing it, so
reading without a limit will cause OutOfMemoryErrors if the file is
sufficiently large.  This is not a security problem (if someone can
trick the user into installing new certificates, crashing the cert
installer dialog is by far the least impactful thing they could do),
but it's nice to not crash.

Bug: 32320490
Test: manual testing
Change-Id: Ib8dbfe06304481fd682297c680841705b8c4ad7c