Install a certificate together with a private key
The CertInstaller supported a feature where a user certificate can be
installed in two steps:
(1) An intent is received with the private&public key pair, the private key
is saved.
(2) An intent is received with the user certificate, the private key
associated with it is loaded and saved with the certificate.

The problem with this flow is that user authentication happens only on
the 2nd phase, allowing an attacker to mess with the keys saved in (1),
potentially disrupting certificate installation flow. CertInstaller
should not be in the business of storing intermediate client data.

Require that the intent sent to install a user certificate contains both
the user certificate and the associated private key. This is less error
prone and prevents an attacker from changing internal CertInstaller
state without user authentication.

Bug: 65543118
Test: Manual (see PoC in the bug).
Change-Id: Ie17505c70adf0c1af748c5aae42b9ba7102aba72