Confirm user credentials _only_ for CA certificate
Confirm the user's credentials only if a CA certificate is installed,
not if a user certificate + public key is installed.

Change Ib998a2fcdff6ceaac515e99245fd5fe3bf202d5a introduced a regression
where the users credentials are confirmed if
CredentialHelper.hasCaCerts() returns true.

However the "CA certificates" in the credentials helper has multiple
uses - in case a client certificate chain is installed it will contain
the certificates that are not the leaf client certificate.
It does not make sense to confirm credentials in this case.

Bug: 134057817
Test: Manual
Change-Id: I5e5de99808be5bc72bf6b3f25860d8ce8df9cd32