KeyChain: Do not allow hiding Cert Install dialog
Do not allow apps to float a window on top of the certificate
installation / naming dialog.

This obscures the CA certificate installation dialog and could be used
to trick a user into installing a CA certificate.

This is fixed by adding the HIDE_NON_SYSTEM_OVERLAY_WINDOWS system
flag when the activity is created (onCreate), so that another activity
starting in the foreground would not be able to obscure the dialog.

Bug: 139017101
Test: Manual, with an app that floats a window.
Change-Id: Iff8e678743c3883cf1f7f64390097a768ca00856